requirements.txt
jieba==0.42.1
requests==2.31.0
streamlit==1.32.1
streamlit-echarts==0.4.0
beautifulsoup4==4.12.3
gensim==4.3.2
snownlp==0.12.3
wordcloud==1.9.3
pillow==10.2.0
matplotlib==3.8.3
numpy==1.26.4
scikit-learn==1.5.0
